+++ app/models/announcement.rb
@@ -1,0 +1,4 @@
+class Announcement < ActiveRecord::Base
+  belongs_to :teacher
+  validates_presence_of :title, :content
+end