+++ app/models/user.rb
@@ -13,6 +13,8 @@ class User < ActiveRecord::Base
 
   has_and_belongs_to_many :class_rooms
 
+  has_many :attachments
+
   def set_password
     if Rails.env.production?
       generated_password = Devise.friendly_token.first(8)
@@ -20,5 +22,9 @@ class User < ActiveRecord::Base
       generated_password = '12341234'
     end
     self.password = generated_password
+  end
+
+  def class_room_user_by_class_room(class_room)
+    ClassRoomsUsers.where("user_id = ? AND class_room_id = ?", self.id, class_room.id).first
   end
 end