+++ spec/views/announcements/index.html.haml_spec.rb
@@ -1,0 +1,26 @@
+require 'spec_helper'
+
+describe "announcements/index" do
+  before(:each) do
+    assign(:announcements, [
+      stub_model(Announcement,
+        :title => "Title",
+        :content => "MyText",
+        :teacher => nil
+      ),
+      stub_model(Announcement,
+        :title => "Title",
+        :content => "MyText",
+        :teacher => nil
+      )
+    ])
+  end
+
+  it "renders a list of announcements" do
+    render
+    # Run the generator again with the --webrat flag if you want to use webrat matchers
+    assert_select "tr>td", :text => "Title".to_s, :count => 2
+    assert_select "tr>td", :text => "MyText".to_s, :count => 2
+    assert_select "tr>td", :text => nil.to_s, :count => 2
+  end
+end