+++ spec/views/announcements/show.html.haml_spec.rb
@@ -1,0 +1,19 @@
+require 'spec_helper'
+
+describe "announcements/show" do
+  before(:each) do
+    @announcement = assign(:announcement, stub_model(Announcement,
+      :title => "Title",
+      :content => "MyText",
+      :teacher => nil
+    ))
+  end
+
+  it "renders attributes in <p>" do
+    render
+    # Run the generator again with the --webrat flag if you want to use webrat matchers
+    rendered.should match(/Title/)
+    rendered.should match(/MyText/)
+    rendered.should match(//)
+  end
+end