+++ spec/views/attachments/index.html.haml_spec.rb
@@ -1,0 +1,26 @@
+require 'spec_helper'
+
+describe "attachments/index" do
+  before(:each) do
+    assign(:attachments, [
+      stub_model(Attachment,
+        :user => nil,
+        :class_room => nil,
+        :title => "Title"
+      ),
+      stub_model(Attachment,
+        :user => nil,
+        :class_room => nil,
+        :title => "Title"
+      )
+    ])
+  end
+
+  it "renders a list of attachments" do
+    render
+    # Run the generator again with the --webrat flag if you want to use webrat matchers
+    assert_select "tr>td", :text => nil.to_s, :count => 2
+    assert_select "tr>td", :text => nil.to_s, :count => 2
+    assert_select "tr>td", :text => "Title".to_s, :count => 2
+  end
+end