+++ spec/views/attachments/show.html.haml_spec.rb
@@ -1,0 +1,19 @@
+require 'spec_helper'
+
+describe "attachments/show" do
+  before(:each) do
+    @attachment = assign(:attachment, stub_model(Attachment,
+      :user => nil,
+      :class_room => nil,
+      :title => "Title"
+    ))
+  end
+
+  it "renders attributes in <p>" do
+    render
+    # Run the generator again with the --webrat flag if you want to use webrat matchers
+    rendered.should match(//)
+    rendered.should match(//)
+    rendered.should match(/Title/)
+  end
+end